refactor: application 분리, lint
