+++ src/main/kotlin/subway/repository/StationRepository.kt
@@ -9,8 +9,6 @@ object StationRepository {
 
     fun add(station: Station) = stations.add(station)
 
-    fun findByName(name: String): Station = stations().first { it.name == name }
-
     fun existsByName(name: String): Boolean = stations().any { it.name == name }
 
     fun deleteByName(name: String) = stations.removeIf { it.name == name }
